Create fizzbuzz.js
HOW IT WORKS

I called the function "fizzBuzz", with a variable of "n", and added a condition. Creating a for loop, which starts at 1, ends when "i" is less than or equal to "n", and increases by increments of 1. This makes "n" a number that will break the loop.

If "i" is evenly divisible by 3 AND if "i" is evenly divisible by 5, then it will log "FizzBuzz" into the console. However, if "i" is only evenly divisible by 3 (but not 5), then it will log "Fizz" into the console.

Else if "i" is only evenly divisible by 5 (but not 3), then it will log "Buzz" into the console. If it is not evenly divisible by both 3 and 5, then "i" will be logged into the console.

Lastly, I called the function with "n" being 100!